move marksegment naming convention to NameUtils.java
