+++ controller/src/main/java/io/pravega/controller/server/bucket/PeriodicWatermarking.java
@@ -33,6 +33,7 @@ import io.pravega.controller.store.stream.records.EpochRecord;
 import io.pravega.controller.store.stream.records.StreamSegmentRecord;
 import io.pravega.controller.store.stream.records.WriterMark;
 import io.pravega.controller.task.Stream.StreamMetadataTasks;
+import io.pravega.shared.NameUtils;
 import io.pravega.shared.segment.StreamSegmentNameUtils;
 import io.pravega.shared.watermarks.SegmentWithRange;
 import io.pravega.shared.watermarks.Watermark;
@@ -151,7 +152,7 @@ public class PeriodicWatermarking {
 
     private CompletableFuture<Void> createMarkStream(String scope, String streamName, WatermarkClient watermarkClient) {
         CompletableFuture<Void> future;
-        future = streamMetadataTasks.createStream(scope, StreamSegmentNameUtils.getMarkSegmentForStream(streamName),
+        future = streamMetadataTasks.createStream(scope, NameUtils.getMarkStreamForStream(streamName),
                 StreamConfiguration.builder().scalingPolicy(ScalingPolicy.fixed(1)).build(), System.currentTimeMillis())
                     .thenAccept(status -> {
                                         switch (status) {
@@ -410,7 +411,7 @@ public class PeriodicWatermarking {
         @VisibleForTesting
         WatermarkClient(Stream stream, SynchronizerClientFactory clientFactory) {
             this.client = clientFactory.createRevisionedStreamClient(
-                    StreamSegmentNameUtils.getMarkSegmentForStream(stream.getStreamName()), 
+                    NameUtils.getMarkStreamForStream(stream.getStreamName()), 
                     new WatermarkSerializer(), SynchronizerConfig.builder().build());
             this.windowStart = new AtomicInteger();
             windowSize = WINDOW_SIZE;
