+++ shared/protocol/src/test/java/io/pravega/shared/segment/StreamSegmentNameUtilsTests.java
@@ -134,12 +134,4 @@ public class StreamSegmentNameUtilsTests {
         AssertExtensions.assertThrows("No tokens supplied", () -> StreamSegmentNameUtils.getQualifiedTableName("scope"), 
                 e -> e instanceof IllegalArgumentException);
     }
-
-    @Test
-    public void testMarkSegmentName() {
-        String myStream = "myStream";
-        String name = StreamSegmentNameUtils.getMarkSegmentForStream(myStream);
-        assertTrue(name.endsWith(myStream));
-        assertTrue(name.startsWith(StreamSegmentNameUtils.getMARK_PREFIX()));
-    }
 }
